feat: Improve accessibility and navigation handling
- Add useClickHandler and useKeyboardHandler hooks - Update navigation handler for better accessibility - Enhance a11y utilities with keyboard constants - Update hook exports
