+++ src/hooks/index.js
@@ -1,3 +1,6 @@
+export * from './useClickHandler';
 export * from './useDeleteHandler';
 export { default as useImageErrorHandler } from './useImageErrorHandler';
+export * from './useKeyboardHandler';
 export * from './useNavigationHandler';
+export * from './useImageLoader';
