+++ src/utils/index.js
@@ -1,2 +1,3 @@
 export * from './a11y';
 export * from './cssConstants';
+export * from './appConstants';
